general: minor code tidy up
