+++ internal/model/args.go
@@ -48,21 +48,5 @@ type RangeReadCloser struct {
 	Closer      io.Closer
 }
 
-type Closers struct {
-	closers []io.Closer
-}
-
-func (c *Closers) Close() (err error) {
-	for _, closer := range c.closers {
-		if closer != nil {
-			_ = closer.Close()
-		}
-	}
-	return nil
-}
-func (c *Closers) Add(closer io.Closer) {
-	c.closers = append(c.closers, closer)
-}
-
 type WriterFunc func(w io.Writer) error
 type RangeReaderFunc func(httpRange http_range.Range) (io.ReadCloser, error)
